Solar event system in place
How it works so far:
Timer shows how long till the next event
Then the event has a duration where the player loses health
Then the whole process repeats after the event is done

What needs to be added/discusses  for this specific system:
Do we want it to be a flat health decrease?
What should be the durations of the whole cycle?
Right now there is no "shade mechanic" Will work on that later